+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/ContentManagerPlugin.java
@@ -16,6 +16,10 @@
  */
 package com.wazuh.contentmanager;
 
+import com.wazuh.contentmanager.client.CommandManagerClient;
+import com.wazuh.contentmanager.model.commandmanager.Command;
+import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
+import org.apache.hc.core5.http.HttpStatus;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.client.Client;
@@ -101,6 +105,29 @@ public class ContentManagerPlugin extends Plugin implements ClusterPlugin, Actio
         ConsumerInfo consumerInfo =
                 Privileged.doPrivilegedRequest(() -> CTIClient.getInstance().getCatalog());
         this.contextIndex.index(consumerInfo);
+        // generate cti command for testing purposes (now method create)
+        String offset = "111";
+        try{
+            String command = Command.create(offset);
+            SimpleHttpResponse response =
+                Privileged.doPrivilegedRequest(() -> CommandManagerClient.getInstance().postCommand(command));
+            switch (response.getCode()){
+                case HttpStatus.SC_OK:
+                    log.info("Received OK response: {}", response.getBody().toString());
+                    break;
+                case HttpStatus.SC_CLIENT_ERROR:
+                    log.error("Client error: {}", response.getBody().toString());
+                    break;
+                case HttpStatus.SC_SERVER_ERROR:
+                    log.error("Internal server error");
+                    break;
+                default:
+                    log.info("Unexpected response code: {}", response.getCode());
+            }
+
+        } catch (IOException e){
+            log.error(e.getMessage());
+        }
     }
 
     /**