+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Third-party integrations maintenance [(#299)](https://github.com/wazuh/wazuh-indexer-plugins/pull/299)
 - Upgrade to Opensearch 2.19.1 [(#304)](https://github.com/wazuh/wazuh-indexer-plugins/pull/304)
 - Improve ECS documentation [(#328)](https://github.com/wazuh/wazuh-indexer-plugins/pull/328)
+- Redefine the command created upon content update [(#339)](https://github.com/wazuh/wazuh-indexer-plugins/pull/339)
 
 
 ### Deprecated